requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.8.0
asyncmy==0.2.10
attrs==25.1.0
blinker==1.9.0
boto3==1.37.4
botocore==1.37.4
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.10
fastapi-mail==1.4.2
frozenlist==1.5.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
h11==0.14.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.5
jmespath==1.0.1
jwt==1.3.1
MarkupSafe==3.0.2
multidict==6.1.0
oauthlib==3.2.2
propcache==0.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pymongo==4.11.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.11.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.0
typing_extensions==4.12.2
urllib3==2.3.0
uuid==1.30
uvicorn==0.34.0
yarl==1.18.3